refactor(chat): extract transcript virtualization controller
